+++ examples/src/main/java/org/multibit/hd/hardware/examples/trezor/usb/TrezorV1WipeAndCreateWalletExample.java
@@ -57,8 +57,8 @@ public class TrezorV1WipeAndCreateWalletExample {
     // Use factory to statically bind the specific hardware wallet
     TrezorV1HidHardwareWallet wallet = HardwareWallets.newUsbInstance(
             TrezorV1HidHardwareWallet.class,
-      Optional.<Short>absent(),
-      Optional.<Short>absent(),
+      Optional.<Integer>absent(),
+      Optional.<Integer>absent(),
       Optional.<String>absent()
     );
