+++ examples/src/main/java/org/multibit/hd/hardware/examples/trezor/usb/TrezorV1WipeAndLoadWalletExample.java
@@ -56,8 +56,8 @@ public class TrezorV1WipeAndLoadWalletExample {
     // Use factory to statically bind the specific hardware wallet
     TrezorV1HidHardwareWallet wallet = HardwareWallets.newUsbInstance(
             TrezorV1HidHardwareWallet.class,
-      Optional.<Short>absent(),
-      Optional.<Short>absent(),
+      Optional.<Integer>absent(),
+      Optional.<Integer>absent(),
       Optional.<String>absent()
     );
